Refactor sript/helper code for better organization
Moves various helper functions from the `scripts` file to
various `helper` classes, which is a more logical organization
for this repo, dividing supporting code from the method in which
it is invoked.

Further the test suite has been updated to reflect this new
structure
