+++ scripts/lambdaRun.py
@@ -2,15 +2,11 @@ import subprocess
 import sys
 import os
 import re
-import yaml
-import json
-
-from collections import ChainMap
 
 from helpers.logHelpers import createLog
 from helpers.errorHelpers import InvalidExecutionType
-from helpers.configHelpers import loadEnvFile
-from helpers.clientHelpers import createAWSClient
+from helpers.clientHelpers import createEventMapping
+from helpers.configHelpers import setEnvVars
 
 logger = createLog('runScripts')
 
@@ -72,101 +68,6 @@ def main():
     else:
         logger.error('Execution type not recognized! {}'.format(runType))
         raise InvalidExecutionType('{} is not a valid command'.format(runType))
-
-
-def loadEnvVars(runType):
-    # Load env variables from relevant .yaml file
-    envDict = loadEnvFile(runType, 'config/{}.yaml')
-
-    # Overwrite/add any vars in the core config.yaml file
-    configDict = loadEnvFile(runType, None)
-
-    combinedConfig = ChainMap(envDict, configDict)
-
-    return combinedConfig
-
-
-def setEnvVars(runType):
-
-    envVars = loadEnvVars(runType)
-
-    try:
-        with open('run_config.yaml', 'w') as newConfig:
-            yaml.dump(
-                dict(envVars),
-                newConfig,
-                default_flow_style=False
-            )
-    except IOError as err:
-        logger.error(('Script lacks necessary permissions, '
-                      'ensure user has permission to write to directory'))
-        raise err
-
-
-def createEventMapping(runType):
-    logger.info('Creating event Source mappings for Lambda')
-    try:
-        with open('config/event_sources_{}.json'.format(runType)) as sources:
-            try:
-                eventMappings = json.load(sources)
-            except json.decoder.JSONDecodeError as err:
-                logger.error('Unable to parse JSON file')
-                raise err
-    except FileNotFoundError:
-        logger.info('No Event Source mapping provided')
-        return
-    except IOError as err:
-        logger.error('Unable to open JSON file')
-        raise err
-
-    if len(eventMappings['EventSourceMappings']) < 1:
-        logger.info('No event sources defined')
-        return
-
-    configDict = loadEnvVars(runType)
-
-    lambdaClient = createAWSClient('lambda', configDict)
-
-    for mapping in eventMappings['EventSourceMappings']:
-        logger.debug('Adding event source mapping for function')
-
-        createKwargs = {
-            'EventSourceArn': mapping['EventSourceArn'],
-            'FunctionName': configDict['function_name'],
-            'Enabled': mapping['Enabled'],
-            'BatchSize': mapping['BatchSize'],
-        }
-
-        if 'StartingPosition' in mapping:
-            createKwargs['StartingPosition'] = mapping['StartingPosition']
-            if mapping['StartingPosition'] == 'AT_TIMESTAMP':
-                createKwargs['StartingPositionTimestamp'] = mapping['StartingPositionTimestamp']  # noqa: E50
-
-        try:
-            lambdaClient.create_event_source_mapping(**createKwargs)
-        except lambdaClient.exceptions.ResourceConflictException as err:
-            logger.info('Event Mapping already exists, update')
-            logger.debug(err)
-            updateEventMapping(lambdaClient, mapping, configDict)
-
-
-def updateEventMapping(client, mapping, configDict):
-
-    listSourceKwargs = {
-        'EventSourceArn': mapping['EventSourceArn'],
-        'FunctionName': configDict['function_name'],
-        'MaxItems': 1
-    }
-    sourceMappings = client.list_event_source_mappings(**listSourceKwargs)
-    mappingMeta = sourceMappings['EventSourceMappings'][0]
-
-    updateKwargs = {
-        'UUID': mappingMeta['UUID'],
-        'FunctionName': configDict['function_name'],
-        'Enabled': mapping['Enabled'],
-        'BatchSize': mapping['BatchSize'],
-    }
-    client.update_event_source_mapping(**updateKwargs)
 
 
 if __name__ == '__main__':
