+++ tests/test_config.py
@@ -1,8 +1,9 @@
 import unittest
 from yaml import YAMLError
-from unittest.mock import patch
+from unittest.mock import patch, mock_open, call
+from collections import ChainMap
 
-from helpers.configHelpers import loadEnvFile
+from helpers.configHelpers import loadEnvFile, setEnvVars, loadEnvVars
 
 
 class TestConfig(unittest.TestCase):
@@ -20,7 +21,7 @@ class TestConfig(unittest.TestCase):
     @patch('yaml.load', side_effect={'testing': True})
     def test_load_env_failure(self, mock_yaml):
         try:
-            resDict = loadEnvFile('development', 'missing/{}.yaml')
+            loadEnvFile('development', 'missing/{}.yaml')
         except FileNotFoundError:
             pass
         self.assertRaises(FileNotFoundError)
@@ -33,7 +34,69 @@ class TestConfig(unittest.TestCase):
     @patch('yaml.load', side_effect=YAMLError)
     def test_read_env_failure(self, mock_yaml):
         try:
-            resDict = loadEnvFile('development', 'config/{}.yaml')
+            loadEnvFile('development', 'config/{}.yaml')
         except YAMLError:
             pass
         self.assertRaises(YAMLError)
+    
+    mockReturns = ChainMap(
+        {
+            'environment_variables': {
+                'test': 'world'
+            }
+        },
+        {}
+    )
+
+    @patch('helpers.configHelpers.loadEnvFile', side_effect=[
+        {'test1': 'hello'},
+        {'test2': 'world'}
+    ])
+    def test_load_env(self, mock_load):
+        testChain = loadEnvVars('test')
+        self.assertIsInstance(testChain, ChainMap)
+        self.assertEqual(testChain['test1'], 'hello')
+
+    @patch('helpers.configHelpers.loadEnvVars', return_value=mockReturns)
+    @patch('builtins.open', new_callable=mock_open, read_data='data')
+    def test_envVar_success(self, mock_file, mock_env):
+        setEnvVars('development')
+        mock_env.assert_has_calls([call('development')])
+
+    mockLoadReturn = ChainMap({
+            'environment_variables': {
+                'test': 'world'
+            }
+        },{
+            'environment_variables': {
+                'jerry': 'hello'
+            }
+        })
+
+    @patch('helpers.configHelpers.loadEnvVars', return_value=mockLoadReturn)
+    def test_envVar_parsing(self, mock_env):
+        m = mock_open()
+        with patch('builtins.open', m, create=True):
+            setEnvVars('development')
+            mock_env.assert_called_once()
+            confHandle = m()
+            confHandle.write.assert_called()
+
+    @patch('helpers.configHelpers.loadEnvVars', return_value=mockLoadReturn)
+    @patch('builtins.open', side_effect=IOError())
+    def test_envVar_permissions(self, mock_file, mock_env):
+        try:
+            setEnvVars('development')
+        except IOError:
+            pass
+        self.assertRaises(IOError)
+
+    mockReturns = ChainMap(
+        {
+            'function_name': 'tester',
+            'environment_variables': {
+                'jerry': 'hello'
+            }
+        },
+        {}
+    )
